# docs/swagger.yaml
basePath: /
definitions:
...
  types.Stat:
    properties:
      dailyAccounts:
        description: new created accounts at this day
        type: integer
      dailyBuckets:
        description: new created buckets at this day
        type: integer
      dailyNeedles:
        description: new created needles at this day
        type: integer
      dailyVolumes:
        description: new created volumes at this day
        type: integer
      day:
        type: string
      totalAccounts:
        description: total accounts until this day
        type: integer
      totalBuckets:
        description: total buckets until this day
        type: integer
      totalNeedles:
        description: total needles until this day
        type: integer
      totalVolumes:
        description: total volumes until this day
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a HUB API.
  title: HUB API
  version: 0.0.1
paths:
...
  /api/stat:
    get:
      consumes:
      - application/json
      description: Get daily statistics for accounts, needles, and volumes using GET
        method
      parameters:
      - description: 'Number of days to return (default: 30)'
        in: query
        name: count
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Array of daily statistics
          schema:
            items:
              $ref: '#/definitions/types.Stat'
            type: array
        "400":
          description: Invalid count parameter
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get statistics by GET
      tags:
      - statistics
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Get daily statistics for accounts, needles, and volumes using POST
        method
      parameters:
      - description: 'Number of days to return (default: 30)'
        in: formData
        name: count
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Array of daily statistics
          schema:
            items:
              $ref: '#/definitions/types.Stat'
            type: array
        "400":
          description: Invalid count parameter
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get statistics by POST
      tags:
      - statistics
swagger: "2.0"